Refactor proxy setting.

+++ cmd/spoof-dpi/main.go
@@ -36,7 +36,7 @@ func main() {
 	}
 
 	if config.SystemProxy {
-		if err := util.SetOsProxy(config.Port); err != nil {
+		if err := util.SetOsProxy(uint16(config.Port)); err != nil {
 			logger.Fatal().Msgf("error while changing proxy settings: %s", err)
 		}
 		defer func() {
